move helper classes in a subdirectory/namespace/file
